requirements.txt
# newui — минимальные зависимости
atlassian-python-api>=3.0.0
requests>=2.28.0
python-dotenv>=1.0.0
customtkinter>=5.0.0

# Опционально для ИИ-агента в UI
# langchain-core
# langgraph
